// app/(authContext)/(platform)/dashboard/kursy/[slug]/strona/types/sections/CourseContentSectionProps.ts
export type CourseContentSectionProps = {
  backgroundColor: string;

  courseContent: {
    title: string;
    titleTextColor: string;
    subtitle: string;
    subtitleTextColor: string;

    sectionTitleBackgroundColor: string;
    sectionTitleTextColor: string;
    sectionArrowTextColor: string;

    stepBackgroundColor: string;
    stepTextColor: string;
    stepIconColor: string;
    stepDurationColor: string;

    sectionBorderColor: string;
    stepsDividerColor: string;
  };
};